requirements.txt
hubspot-api-client
python-dotenv
oauth2client
gspread